Todomanía/requirements.txt
pillow
django-crispy-forms
